+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KKGJ3CD2TX1VWEM03DKZR2KW"
+        "01KKGT4G8AYW22EMMWMW7AJVPB"
     )
     r"""The unique identifier of the data source"""