+++ packages/orq-rc/src/orq_ai_sdk/models/listidentitiesop.py
@@ -208,7 +208,7 @@ class ListIdentitiesData(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-03-12T08:19:52.221Z")
+    updated: Optional[datetime] = parse_datetime("2026-03-12T10:40:16.959Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")