+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KKGJ3CD40HTSTQP2K4ACXZEP"
+        "01KKGT4G8BRS50JV31AJ0RCCAJ"
     )
     r"""The unique identifier of the data source"""